utils move to make them local (for making resources a package that moves to base)
