+++ aws_topology/stackstate_checks/aws_topology/resources/autoscaling.py
@@ -1,4 +1,4 @@
-from ..utils import make_valid_data
+from .utils import make_valid_data
 from .registry import RegisteredResourceCollector
