+++ aws_topology/stackstate_checks/aws_topology/resources/ecs.py
@@ -1,6 +1,6 @@
 import boto3
 from botocore.config import Config
-from ..utils import make_valid_data, with_dimensions
+from .utils import make_valid_data, with_dimensions
 from .registry import RegisteredResourceCollector
